[TASK] Allow additional arguments in UriBuilder
This change add a new method UriBuilder::setAdditionalArguments. Those
arguments will be add to the request URI but are not used in the route
resolving logic. This can be useful per example for search result
facetting where a lots of links are created and we can face cache
flooding issues.